advisor_service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
Pillow==10.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
celery==5.3.1
redis==5.0.0
joblib==1.3.1
tensorflow==2.15.0
